feat(cqcMainServiceApprovalAndRejectionButtons): Move approval-or-rejection-dialog component into new components folder inside admin
